Follow the example/* syntax for the new pub site.
